Export function to prompt user to select/create connection
