+++ src/objectExplorer/treeNodeInfo.ts
@@ -6,9 +6,9 @@
 import * as vscode from 'vscode';
 import { NodeInfo } from '../models/contracts/objectExplorer/nodeInfo';
 import { ObjectExplorerUtils } from './objectExplorerUtils';
-import { IConnectionCredentials } from '../models/interfaces';
 import Constants = require('../constants/constants');
 import { ObjectMetadata } from '../models/contracts/metadata/metadataRequest';
+import { IConnectionCredentials } from 'vscode-mssql';
 
 export class TreeNodeInfo extends vscode.TreeItem {
