+++ test/connectionCredentials.test.ts
@@ -16,10 +16,11 @@ import { ConnectionStore } from '../src/models/connectionStore';
 import { ConnectionCredentials } from '../src/models/connectionCredentials';
 import { IPrompter, IQuestion} from '../src/prompts/question';
 import { TestPrompter } from './stubs';
-import { IConnectionProfile, IConnectionCredentials } from '../src/models/interfaces';
+import { IConnectionProfile } from '../src/models/interfaces';
 import VscodeWrapper from '../src/controllers/vscodeWrapper';
 
 import assert = require('assert');
+import { IConnectionCredentials } from 'vscode-mssql';
 
 suite('ConnectionCredentials Tests', () => {
     let defaultProfile: interfaces.IConnectionProfile;
